+++ loop/sfm/lightgbm/__init__.py
@@ -4,25 +4,8 @@ import loop.sfm.lightgbm.tradeable_regressor as tradeable_regressor
 import loop.sfm.lightgbm.tradeline_multiclass as tradeline_multiclass
 import loop.sfm.lightgbm.utils as utils
 
-# Import mega model utilities
-from .utils import (
-    MegaModelDataSampler,
-    run_mega_model_experiment,
-    run_enhanced_megamodel_with_uel,
-    integrate_enhanced_megamodel_into_workflow,
-    get_best_model_from_results,
-    predict_with_best_model,
-    save_experiment_results
-)
 
 __all__ = [
-    'MegaModelDataSampler',
-    'run_mega_model_experiment',
-    'run_enhanced_megamodel_with_uel',
-    'integrate_enhanced_megamodel_into_workflow',
-    'get_best_model_from_results',
-    'predict_with_best_model',
-    'save_experiment_results',
     'tradeable_regressor',
     'tradeline_multiclass',
     'utils'